+++ pkg/assert/regexp.go
@@ -10,24 +10,18 @@ func regexpMatch(text string, expr string) bool {
 	return r.FindStringIndex(text) != nil
 }
 
-func Regexp(text, expr interface{}) (bool, string) {
+func Matches(text, expr interface{}) (bool, string) {
 	if test(text, expr, regexpMatch) {
 		return true, "OK"
 	}
 
-	// if regexpMatch(expr, text) {
-	// 	return true, "OK"
-	// }
-	return false, fmt.Sprintf("regexp | `%v` should match `%v`", prettyLine(text), expr)
+	return false, fmt.Sprintf("matches | `%v` should match `%v`", prettyLine(text), expr)
 }
 
-func NotRegexp(text, expr interface{}) (bool, string) {
+func NotMatches(text, expr interface{}) (bool, string) {
 	if !test(text, expr, regexpMatch) {
 		return true, "OK"
 	}
-	// if !regexpMatch(expr, text) {
-	// 	return true, "OK"
-	// }
 
-	return false, fmt.Sprintf("not_regexp | `%v` should not match `%v`", prettyLine(text), expr)
+	return false, fmt.Sprintf("not_matches | `%v` should not match `%v`", prettyLine(text), expr)
 }