+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Add
 
-- assert: header_exists/cookie_exists/body_icontains/body_regexp/body_not_regexp
+- assert: header_exists/cookie_exists/body_icontains/body_matches/body_not_matches
 
 ### Documentation